requirements.txt
# OddsPortal Scraper Dependencies
playwright>=1.40.0      # Browser automation (handles JS-rendered content)
requests>=2.31.0       # HTTP requests for potential API calls
beautifulsoup4>=4.12.0 # HTML parsing
pandas>=2.0.0          # Data handling and export
python-dotenv>=1.0.0   # Optional: env config
openpyxl>=3.1.0       # Excel export
